+++ is_shoppingcart.php
@@ -11,6 +11,8 @@ if (file_exists(__DIR__ . '/vendor/autoload.php')) {
 }
 
 use Oksydan\IsShoppingcart\Configuration\ShoppingCartConfiguration;
+use Oksydan\IsShoppingcart\Configuration\PreviewTypes;
+use Oksydan\IsShoppingcart\Configuration\NotificationsTypes;
 use Oksydan\IsShoppingcart\Hook\HookInterface;
 use PrestaShop\PrestaShop\Adapter\SymfonyContainer;
 use Symfony\Component\DependencyInjection\Exception\ServiceNotFoundException;
@@ -21,7 +23,7 @@ class Is_shoppingcart extends Module
     {
         $this->name = 'is_shoppingcart';
         $this->tab = 'front_office_features';
-        $this->version = '3.0.1';
+        $this->version = '4.0.0';
         $this->author = 'Igor Stępień';
         $this->need_instance = 0;
 
@@ -44,6 +46,10 @@ class Is_shoppingcart extends Module
      */
     public function install(): bool
     {
+        if (Shop::isFeatureActive()) {
+            Shop::setContext(Shop::CONTEXT_ALL);
+        }
+
         return parent::install()
             && $this->installHooks()
             && $this->installConfiguration();
@@ -63,7 +69,8 @@ class Is_shoppingcart extends Module
      */
     public function installConfiguration(): bool
     {
-        return \Configuration::updateValue(ShoppingCartConfiguration::IS_BLOCK_CART_AJAX, 1);
+        return \Configuration::updateGlobalValue(ShoppingCartConfiguration::IS_CART_NOTIFICATION_TYPE, NotificationsTypes::NOTIFICATION_TYPE_TOAST) &&
+            \Configuration::updateGlobalValue(ShoppingCartConfiguration::IS_CART_PREVIEW_TYPE, PreviewTypes::PREVIEW_TYPE_OFFCANVAS);
     }
 
     /**
